requirements.txt
fontawesome-free==5.15.2
django-phonenumber-field[phonenumbers]
